# docs/BeneficiaryAPI.Swagger20.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Beneficiary API",
    "contact": {}
  },
  "host": "app:3000",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/bank": {
      "post": {
        "summary": "Create Bank",
        "tags": ["bank"],
        "operationId": "CreateBank",
        "deprecated": false,
        "produces": ["text/plain; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateBankRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "string"
            },
            "examples": {
              "text/plain; charset=utf-8": "Created"
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "7"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"7-rM9AyJuqT6iOan/xHh+AW+7K/T8\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 13 Jun 2021 14:55:31 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
    "/banks": {
      "get": {
        "summary": "Get banks",
        "tags": ["bank"],
        "operationId": "Getbanks",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/bank/{id}": {
      "get": {
        "summary": "Get bank",
        "tags": ["bank"],
        "operationId": "Getbank",
        "deprecated": false,
        "produces": ["application/json; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Getbank"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "id": "BANCO_DO_BRASIL",
                "name": "Banco do Brasil (001)"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "55"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"37-62S4geMD88SzlDOzhmLt/6CXp5I\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 13 Jun 2021 14:59:56 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Bank",
        "tags": ["bank"],
        "operationId": "UpdateBank",
        "deprecated": false,
        "produces": ["text/plain; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateBankRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            },
            "examples": {
              "text/plain; charset=utf-8": "OK"
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "2"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"2-nOO9QiTIwXgNtWtBJezz8kv3SLc\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 13 Jun 2021 15:01:47 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Bank",
        "tags": ["bank"],
        "operationId": "DeleteBank",
        "deprecated": false,
        "produces": ["text/plain; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            },
            "examples": {
              "text/plain; charset=utf-8": "OK"
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "2"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"2-nOO9QiTIwXgNtWtBJezz8kv3SLc\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 13 Jun 2021 15:03:41 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
    "/bank/{id}/accountTypes": {
      "get": {
        "summary": "Get account types",
        "tags": ["bank"],
        "operationId": "Getaccounttypes",
        "deprecated": false,
        "produces": ["application/json; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Getaccounttype"
              }
            },
            "examples": {
              "application/json; charset=utf-8": [
                {
                  "id": 1,
                  "type": "CONTA_CORRENTE",
                  "name": "Conta corrente"
                }
              ]
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "58"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"3a-d6Ft+Ym1pm62zkh3P2w5ElHDArI\""
              },
              "Date": {
                "type": "string",
                "default": "Tue, 15 Jun 2021 01:20:54 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "CreateBankRequest": {
      "title": "CreateBankRequest",
      "example": {
        "id": "BANCO_DO_BRASIL",
        "name": "Banco do Brasil (001)"
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "required": ["id", "name"]
    },
    "Getbank": {
      "title": "Getbank",
      "example": {
        "id": "BANCO_DO_BRASIL",
        "name": "Banco do Brasil (001)"
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "required": ["id", "name"]
    },
    "UpdateBankRequest": {
      "title": "UpdateBankRequest",
      "example": {
        "name": "Banco do Brasil"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      },
      "required": ["name"]
    },
    "Getaccounttype": {
      "title": "Getaccounttype",
      "example": {
        "id": 1,
        "type": "CONTA_CORRENTE",
        "name": "Conta corrente"
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "required": ["id", "type", "name"]
    },
...
  },
...
}